requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
bcrypt==4.0.1
click==8.1.7
ecdsa==0.18.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
httptools==0.6.1
iniconfig==2.0.0
Mako==1.2.4
passlib==1.7.4
pluggy==1.3.0
psycopg==3.1.12
psycopg-pool==3.1.8
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyasn1==0.5.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pyscopg2==66.0.2
pytest==7.4.3
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
SQLAlchemy==2.0.23
starlette==0.27.0
uvicorn==0.23.2
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
